tests/pyproject.toml
[project]
name = "dnd-master-tests"
version = "1.0.0"
description = "Test suite for DnD Master application"
requires-python = ">=3.12"
dependencies = [
    # Testing framework
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",

    # HTTP testing
    "requests>=2.32.5",
    "httpx==0.27.0",

    # FastAPI and dependencies
    "fastapi==0.115.5",
    "uvicorn==0.32.1",
    "pydantic>=2.7.4",

    # Google Cloud AI Platform
    "google-cloud-aiplatform==1.71.1",
    "google-genai==0.2.2",

    # ChromaDB for RAG tests
    "chromadb==0.5.23",

    # Langchain for rule agent
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-text-splitters>=0.3.0",

    # OpenAI
    "openai==1.57.2",

    # Data processing (needed by backend)
    "pandas==2.2.3",
    "numpy>=2.1.0",
    
    # Code quality tools (for CI/CD)
    "black>=24.0.0",
    "flake8>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Minimal package definition to satisfy hatchling
packages = ["dnd_master_tests"]
